Fixed issue with coercing photo result list → data.frame
Notes: rbind and rbind.fill strip out NULLs. They generally expect equal length nested lists to make this work. In this case empty fields are NULL (which is good), but that results in unequal length vectors.
The resulting data.frame then contains lists (legal in R) rather than vectors. This is problematic down the line.
This issue is now fixed with a helper function (rbindfillnull) that replaces NULL →  NA so that a coercion is possible.
This will carry over to other functions (observations, checlists as well). Will look for a better solution.